v1.0.2　vaultの不具合修正

+++ src/main/java/red/man10/man10commandlibrary/VaultManager.java
@@ -51,7 +51,7 @@ public class VaultManager {
     public void showBalance(UUID uuid){
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         double money = getBalance(uuid);
-        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + (int)money);
+        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + money);
     }
     /////////////////////////////////////
     //      引き出し
@@ -65,7 +65,7 @@ public class VaultManager {
         EconomyResponse resp = economy.withdrawPlayer(p,money);
         if(resp.transactionSuccess()){
             if(p.isOnline()) {
-                p.getPlayer().sendMessage(ChatColor.YELLOW + "$" + (int) money + "支払いました");
+                p.getPlayer().sendMessage(ChatColor.YELLOW + "$" + money + "支払いました");
             }
             return true;
         }
@@ -84,7 +84,7 @@ public class VaultManager {
         EconomyResponse resp = economy.depositPlayer(p,money);
         if(resp.transactionSuccess()){
             if(p.isOnline()){
-                p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+(int)money+"受取りました");
+                p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+money+"受取りました");
             }
             return true;
         }
